+++ packages/article/article-body/article-body-paragraph.web.js
@@ -1,7 +1,10 @@
 import React from "react";
 import PropTypes from "prop-types";
-import { ParagraphContainer, Paragraph } from "./styles/body/responsive";
-import styles from "./styles/body";
+import {
+  ParagraphContainer,
+  Paragraph
+} from "../styles/article-body/responsive";
+import styles from "../styles/article-body";
 
 const BodyParagraph = props => (
   <ParagraphContainer