+++ packages/article/styles/article-meta/shared.js
@@ -20,14 +20,14 @@ const sharedStyles = {
   datePublication: {
     color: "#696969",
     fontSize: 13,
-    lineHeight: 13,
+    lineHeight: 15,
     fontFamily: "GillSansMTStd-Medium"
   },
   byline: {
     fontSize: 13,
     fontFamily: "GillSansMTStd-Medium",
     color: "#696969",
-    lineHeight: 13
+    lineHeight: 15
   }
 };