+++ packages/jest-configurator/setup.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
 set -e
 
-mkdir fixtures/all/node_modules
-mkdir fixtures/ignore/node_modules
+mkdir -p fixtures/all/node_modules
+mkdir -p fixtures/ignore/node_modules