+++ packages/ad/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.6.10"></a>
+## [0.6.10](https://github.com/newsuk/times-components/compare/@times-components/ad@0.6.9...@times-components/ad@0.6.10) (2018-01-18)
+
+
+
+
+**Note:** Version bump only for package @times-components/ad
+
 <a name="0.6.9"></a>
 ## [0.6.9](https://github.com/newsuk/times-components/compare/@times-components/ad@0.6.8...@times-components/ad@0.6.9) (2018-01-16)