+++ packages/gradient/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.11"></a>
+## [0.3.11](https://github.com/newsuk/times-components/compare/@times-components/gradient@0.3.10...@times-components/gradient@0.3.11) (2018-01-18)
+
+
+
+
+**Note:** Version bump only for package @times-components/gradient
+
 <a name="0.3.10"></a>
 ## [0.3.10](https://github.com/newsuk/times-components/compare/@times-components/gradient@0.3.9...@times-components/gradient@0.3.10) (2018-01-16)