+++ packages/responsive-styles/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.10"></a>
+## [0.2.10](https://github.com/newsuk/times-components/compare/@times-components/responsive-styles@0.2.9...@times-components/responsive-styles@0.2.10) (2018-01-17)
+
+
+
+
+**Note:** Version bump only for package @times-components/responsive-styles
+
 <a name="0.2.9"></a>
 ## [0.2.9](https://github.com/newsuk/times-components/compare/@times-components/responsive-styles@0.2.8...@times-components/responsive-styles@0.2.9) (2018-01-16)